Extract Prompt Studio extension runtime
